refactor: 카드 포스트 컴포넌트 분리
